requirements.txt
fritzconnection==1.2.1
pytelegraf==0.3.3
